Enhancement: Merge Provider\Commit into Repository\Commit
