+++ src/Repository/Commit.php
@@ -4,20 +4,92 @@ namespace Localheinz\ChangeLog\Repository;
 
 use Github\Api;
 use Localheinz\ChangeLog\Entity;
+use Localheinz\ChangeLog\Provider\ItemProvider;
 
-class Commit
+class Commit implements ItemProvider
 {
     /**
      * @var Api\Repository\Commits
      */
-    private $commitApi;
+    private $api;
 
     /**
-     * @param Api\Repository\Commits $commitApi
+     * @param Api\Repository\Commits $api
      */
-    public function __construct(Api\Repository\Commits $commitApi)
+    public function __construct(Api\Repository\Commits $api)
     {
-        $this->commitApi = $commitApi;
+        $this->api = $api;
+    }
+
+    /**
+     * @param string $vendor
+     * @param string $package
+     * @param string $startReference
+     * @param string $endReference
+     * @return Entity\Commit[]
+     */
+    public function items($vendor, $package, $startReference, $endReference)
+    {
+        if ($startReference === $endReference) {
+            return [];
+        }
+
+        $start = $this->show(
+            $vendor,
+            $package,
+            $startReference
+        );
+
+        if (null === $start) {
+            return [];
+        }
+
+        $end = $this->show(
+            $vendor,
+            $package,
+            $endReference
+        );
+
+        if (null === $end) {
+            return [];
+        }
+
+        $commits = $this->all($vendor, $package, [
+            'sha' => $start->sha(),
+        ]);
+
+        if (!is_array($commits)) {
+            return [];
+        }
+
+        $range = [];
+
+        $currentStart = $start;
+
+        while (count($commits)) {
+            /* @var Entity\Commit $commit */
+            $commit = array_shift($commits);
+
+            if ($commit->sha() === $currentStart->sha()) {
+                continue;
+            }
+
+            array_push($range, $commit);
+
+            if ($commit->sha() === $end->sha()) {
+                break;
+            }
+
+            if (!count($commits)) {
+                $currentStart = $commit;
+
+                $commits = $this->all($vendor, $package, [
+                    'sha' => $currentStart->sha(),
+                ]);
+            }
+        }
+
+        return $range;
     }
 
     /**
@@ -28,7 +100,7 @@ class Commit
      */
     public function show($vendor, $package, $sha)
     {
-        $response = $this->commitApi->show(
+        $response = $this->api->show(
             $vendor,
             $package,
             $sha
@@ -52,7 +124,7 @@ class Commit
      */
     public function all($vendor, $package, array $params = [])
     {
-        $response = $this->commitApi->all(
+        $response = $this->api->all(
             $vendor,
             $package,
             $params
