+++ tests/Repository/CommitTest.php
@@ -5,15 +5,23 @@ namespace Localheinz\ChangeLog\Test\Repository;
 use Faker;
 use Github\Api;
 use Localheinz\ChangeLog\Entity;
+use Localheinz\ChangeLog\Provider;
 use Localheinz\ChangeLog\Repository;
-use Localheinz\ChangeLog\Test\Util\DataProviderTrait;
+use Localheinz\ChangeLog\Test\Util\FakerTrait;
 use PHPUnit_Framework_MockObject_MockObject;
 use PHPUnit_Framework_TestCase;
 use stdClass;
 
 class CommitTest extends PHPUnit_Framework_TestCase
 {
-    use DataProviderTrait;
+    use FakerTrait;
+
+    public function testImplementsProvidesItemsInterface()
+    {
+        $repository = new Repository\Commit($this->api());
+
+        $this->assertInstanceOf(Provider\ItemProvider::class, $repository);
+    }
 
     public function testShowReturnsCommitEntityWithShaAndMessageOnSuccess()
     {
@@ -21,9 +29,9 @@ class CommitTest extends PHPUnit_Framework_TestCase
         $package = 'bar';
         $sha = 'ad77125';
 
-        $api = $this->commitApi();
+        $api = $this->api();
 
-        $expected = $this->commitData();
+        $expected = $this->commit();
 
         $api
             ->expects($this->once())
@@ -56,7 +64,7 @@ class CommitTest extends PHPUnit_Framework_TestCase
         $package = 'bar';
         $sha = 'ad77125';
 
-        $api = $this->commitApi();
+        $api = $this->api();
 
         $api
             ->expects($this->once())
@@ -86,7 +94,7 @@ class CommitTest extends PHPUnit_Framework_TestCase
         $package = 'bar';
         $sha = 'ad77125';
 
-        $api = $this->commitApi();
+        $api = $this->api();
 
         $api
             ->expects($this->once())
@@ -101,9 +109,9 @@ class CommitTest extends PHPUnit_Framework_TestCase
             ->willReturn('snafu')
         ;
 
-        $commitRepository = new Repository\Commit($api);
+        $repository = new Repository\Commit($api);
 
-        $commits = $commitRepository->all($vendor, $package, [
+        $commits = $repository->all($vendor, $package, [
             'sha' => $sha,
         ]);
 
@@ -116,11 +124,11 @@ class CommitTest extends PHPUnit_Framework_TestCase
         $package = 'bar';
         $sha = 'ad77125';
 
-        $api = $this->commitApi();
+        $api = $this->api();
 
         $expectedCommits = [];
         for ($i = 0; $i < 15; $i++) {
-            array_push($expectedCommits, $this->commitData());
+            array_push($expectedCommits, $this->commit());
         }
 
         $api
@@ -136,9 +144,9 @@ class CommitTest extends PHPUnit_Framework_TestCase
             ->willReturn($this->responseFromCommits($expectedCommits))
         ;
 
-        $commitRepository = new Repository\Commit($api);
+        $repository = new Repository\Commit($api);
 
-        $commits = $commitRepository->all($vendor, $package, [
+        $commits = $repository->all($vendor, $package, [
             'sha' => $sha,
         ]);
 
@@ -153,12 +161,377 @@ class CommitTest extends PHPUnit_Framework_TestCase
         }
     }
 
+    /* HELLO */
+
+    public function testItemsDoesNotFetchCommitsIfStartAndEndReferencesAreTheSame()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = $startReference;
+
+        $api = $this->api();
+
+        $api
+            ->expects($this->never())
+            ->method($this->anything())
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $commits = $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+
+        $this->assertSame([], $commits);
+    }
+
+    public function testItemsDoesNotFetchCommitsIfStartCommitCouldNotBeFound()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = '7fc1c4f';
+
+        $api = $this->api();
+
+        $api
+            ->expects($this->at(0))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($startReference)
+            )
+            ->willReturn(null)
+        ;
+
+        $api
+            ->expects($this->never())
+            ->method('all')
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $commits = $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+
+        $this->assertSame([], $commits);
+    }
+
+    public function testItemsDoesNotFetchCommitsIfEndCommitCouldNotBeFound()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = '7fc1c4f';
+
+        $api = $this->api();
+
+        $api
+            ->expects($this->at(0))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($startReference)
+            )
+            ->willReturn($this->responseFromCommit($this->commit()))
+        ;
+
+        $api
+            ->expects($this->at(1))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($endReference)
+            )
+            ->willReturn(null)
+        ;
+
+        $api
+            ->expects($this->never())
+            ->method('all')
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $commits = $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+
+        $this->assertSame([], $commits);
+    }
+
+    public function testItemsFetchesCommitsUsingShaFromStartCommit()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = '7fc1c4f';
+
+        $api = $this->api();
+
+        $startCommit = $this->commit();
+
+        $api
+            ->expects($this->at(0))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($startReference)
+            )
+            ->willReturn($this->responseFromCommit($startCommit))
+        ;
+
+        $api
+            ->expects($this->at(1))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($endReference)
+            )
+            ->willReturn($this->responseFromCommit($this->commit()))
+        ;
+
+        $api
+            ->expects($this->once())
+            ->method('all')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo([
+                    'sha' => $startCommit->sha,
+                ])
+            )
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+    }
+
+    public function testItemsReturnsArrayOfCommitsFromStartToEndExcludingStart()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = '7fc1c4f';
+
+        $api = $this->api();
+
+        $startCommit = $this->commit();
+
+        $api
+            ->expects($this->at(0))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($startReference)
+            )
+            ->willReturn($this->responseFromCommit($startCommit))
+        ;
+
+        $endCommit = $this->commit();
+
+        $api
+            ->expects($this->at(1))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($endReference)
+            )
+            ->willReturn($this->responseFromCommit($endCommit))
+        ;
+
+        $countBetween = 13;
+        $countAfter = 17;
+
+        $allCommits = array_merge(
+            [
+                $startCommit,
+            ],
+            $this->commits($countBetween),
+            [
+                $endCommit,
+            ],
+            $this->commits($countAfter)
+        );
+
+        $expectedCommits = array_slice(
+            $allCommits,
+            1,
+            $countBetween + 1
+        );
+
+        $api
+            ->expects($this->once())
+            ->method('all')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo([
+                    'sha' => $startCommit->sha,
+                ])
+            )
+            ->willReturn($this->responseFromCommits($allCommits))
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $commits = $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+
+        $this->assertCount(count($expectedCommits), $commits);
+
+        foreach ($commits as $commit) {
+            $expectedCommit = array_shift($expectedCommits);
+
+            $this->assertInstanceOf(Entity\Commit::class, $commit);
+            $this->assertSame($expectedCommit->sha, $commit->sha());
+            $this->assertSame($expectedCommit->message, $commit->message());
+        }
+    }
+
+    public function testItemsFetchesMoreCommitsIfEndIsNotContainedInFirstBatch()
+    {
+        $vendor = 'foo';
+        $package = 'bar';
+        $startReference = 'ad77125';
+        $endReference = '7fc1c4f';
+
+        $api = $this->api();
+
+        $startCommit = $this->commit();
+
+        $api
+            ->expects($this->at(0))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($startReference)
+            )
+            ->willReturn($this->responseFromCommit($startCommit))
+        ;
+
+        $endCommit = $this->commit();
+
+        $api
+            ->expects($this->at(1))
+            ->method('show')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo($endReference)
+            )
+            ->willReturn($this->responseFromCommit($endCommit))
+        ;
+
+        $firstBatch = array_merge(
+            [
+                $startCommit,
+            ],
+            $this->commits(50)
+        );
+
+        $lastCommitFromFirstBatch = end($firstBatch);
+
+        $secondBatch = array_merge(
+            [
+                $lastCommitFromFirstBatch,
+            ],
+            $this->commits(20)
+        );
+
+        $expectedCommits = array_merge(
+            array_slice(
+                $firstBatch,
+                1
+            ),
+            array_slice(
+                $secondBatch,
+                1
+            )
+        );
+
+        $api
+            ->expects($this->at(2))
+            ->method('all')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo([
+                    'sha' => $startCommit->sha,
+                ])
+            )
+            ->willReturn($this->responseFromCommits($firstBatch))
+        ;
+
+        $api
+            ->expects($this->at(3))
+            ->method('all')
+            ->with(
+                $this->equalTo($vendor),
+                $this->equalTo($package),
+                $this->equalTo([
+                    'sha' => $lastCommitFromFirstBatch->sha,
+                ])
+            )
+            ->willReturn($this->responseFromCommits($secondBatch))
+        ;
+
+        $repository = new Repository\Commit($api);
+
+        $commits = $repository->items(
+            $vendor,
+            $package,
+            $startReference,
+            $endReference
+        );
+
+        $this->assertCount(count($expectedCommits), $commits);
+
+        foreach ($commits as $commit) {
+            $expectedCommit = array_shift($expectedCommits);
+
+            $this->assertInstanceOf(Entity\Commit::class, $commit);
+
+            $this->assertSame($expectedCommit->sha, $commit->sha());
+            $this->assertSame($expectedCommit->message, $commit->message());
+        }
+    }
+
     /**
      * @param string $sha
      * @param string $message
      * @return stdClass
      */
-    private function commitData($sha = null, $message = null)
+    private function commit($sha = null, $message = null)
     {
         $data = new stdClass();
 
@@ -166,6 +539,21 @@ class CommitTest extends PHPUnit_Framework_TestCase
         $data->message = $message ?: $this->faker()->unique()->sentence();
 
         return $data;
+    }
+
+    /**
+     * @param int $count
+     * @return stdClass[]
+     */
+    private function commits($count)
+    {
+        $commits = [];
+
+        for ($i = 0; $i < $count; $i++) {
+            array_push($commits, $this->commit());
+        }
+
+        return $commits;
     }
 
     /**
@@ -212,7 +600,7 @@ class CommitTest extends PHPUnit_Framework_TestCase
     /**
      * @return PHPUnit_Framework_MockObject_MockObject
      */
-    private function commitApi()
+    private function api()
     {
         return $this->getMockBuilder(Api\Repository\Commits::class)
             ->disableOriginalConstructor()
